chore: changes the package group to prevent ambiguities
so, apparently there is a br.com.uol.pagseguro:4.3.2 on Maven Central (https://search.maven.org/artifact/br.com.uol.pagseguro/pagseguro-api)
a version that, mind you, went undocumented. So when we published our current version on GitHub Packages (4.2.0 at the time), we weren't
downloading the right version. long story short, that resulted in a LOT of confusion for me: the the code downloaded from Maven and the code
in my computer were different and a bug that shouldn't exist anymore, did. So to avoid this sort of confusion from now on, I'm going to bump
the version again to 4.4.0 and change the package group once and for all

and we lived happily ever after :)